+++ pyccel/stdlib/cwrapper/cwrapper.c
@@ -177,11 +177,7 @@ void capsule_cleanup(PyObject *capsule) {
     free(memory);
 }
 
-#if defined(WIN32) && (PyArray_RUNTIME_VERSION >= NPY_2_0_API_VERSION)
 PyObject* to_pyarray(int nd, enum NPY_TYPES typenum, void* data, int32_t shape[], bool c_order, bool release_memory)
-#else
-PyObject* to_pyarray(int nd, enum NPY_TYPES typenum, void* data, int64_t shape[], bool c_order, bool release_memory)
-#endif
 {
     int FLAGS;
     if (nd == 1) {