+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 -   #1720 : Add support for `Ellipsis` as the only index for an array.
--   #1787 : Ensure STC is installed with Pyccel.
+-   #1787 : Ensure STC v5.0 is installed with Pyccel.
 -   #1656 : Ensure gFTL is installed with Pyccel.
 -   #1694 : Add Python support for list method `extend()`.
 -   #1700 : Add Python support for list method `sort()`.
@@ -48,8 +48,9 @@ All notable changes to this project will be documented in this file.
 -   #1745 : Add support for set method `intersection_update()`.
 -   #2059 : Add C support for returning dictionaries from functions.
 -   #2164 : Add support for dict indexing.
--   #1884 : Add support for dict method `items()`.
 -   #1880 : Add support for dict method `clear()`.
+-   #1884 : Add support for dict method `items()`.
+-   #1884 : Add support for dict method `keys()`.
 -   #1886 : Add Python and C support for dict method `pop()`.
 -   #1936 : Add missing C output for inline decorator example in documentation
 -   #1937 : Optimise `pyccel.ast.basic.PyccelAstNode.substitute` method.